Alterando a estrutura das rotas.
